Implemented Trading
Players are now able to trade moves.

The Trade class contains an Acccept and an Reject method, that will handle the logistics of switching the moves.

The a trade is crated, the move is removed from the offering player, to prevent the player to offer the same move to multiple other players